🐳 feat(stockdb): fixing review comments
Updates the base Docker image from `python:3.13-slim` to `python:3.13-slim-bookworm`. This change
ensures the latest security updates and bug fixes are included in the base image, improving the
overall security and stability of the application.

Additionally, the `apt` commands have been optimized to reduce the number of layers in the final
Docker image, improving build times and reducing the overall image size.